Revert "[Merge] From Unstable"
This reverts commit d1e77d2ac4dd162ec8110153ebc1d1e54b0c5c79.

+++ MMOCoreORB/src/server/zone/objects/creature/commands/HealActionSelf2Command.h
@@ -53,7 +53,9 @@ public:
 
 		uint32 actionHealed = creature->healDamage(creature, CreatureAttribute::ACTION, healAmount);
 
-		forceCostDeducted = (int) MIN(((float) actionHealed / 15.f), forceCost);
+
+		forceCostDeducted = MIN((actionHealed / 15), forceCost);
+
 
 		// Send system message(s).
